Refactor: Update import path for GitHub authorization action
